// ui/src/orval/models/navigationTreeSchema.ts
/**
 * Generated by orval v7.10.0 🍺
 * Do not edit manually.
 * UI Router API
 * Defines the specification for the UI Catalog API
 * OpenAPI spec version: 1.0.2
 */
import type { NavigationTreeTable } from './navigationTreeTable';

export interface NavigationTreeSchema {
  name: string;
  tables: NavigationTreeTable[];
  views: NavigationTreeTable[];
}